requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyMySQL==1.1.0
cryptography==41.0.7
itsdangerous==2.1.2
werkzeug==3.0.1
